Add files via upload
A wellness and fitness promotion website that calculates users BMI, gives health advice based on the results and sets a dietary plan and exercise regimen to improve their health status. Built with PHP, HTML, CSS, and MySQL, the website is very responsive and have been rigorously tested to improve efficiency. Go ahead and check it out!